inline event too, refactor names
