+++ gradle.properties
@@ -13,4 +13,4 @@ junitVersion=5.1.0
 systemProp.jdk.tls.client.protocols="TLSv1,TLSv1.1,TLSv1.2"
 
 # https://jitpack.io/#alfio-event/alf.io-public-frontend -> go to commit tab, set the version
-alfioPublicFrontendVersion=3e5fbb417c
+alfioPublicFrontendVersion=c566584f29
